move dropdown handling to custom hooks
