+++ src/components/App/hooks/index.tsx
@@ -1,4 +1,5 @@
-import useRepos from "./useRepos"
+import useDropdown from "./useDropdown"
 import useFilteredRepos from "./useFilteredRepos"
+import useRepos from "./useRepos"
 
-export { useRepos, useFilteredRepos }
+export { useDropdown, useFilteredRepos, useRepos }
